Pushing Working CSS and some comments
This is the site navbar I have working (though it's a pain in the ass
trying to merge jsf, javascript, xhtml, and html). Should work
everywhere on the site, I tested every possible path. I have some css
for making the site itself look a little prettier, but it crashes the
server, lol.

Noticed when I was testing things for the CSS that a short password will
cause permanant authentication failure, so we may want to force the user
to have a password of a certain length before they can submit the form.
I also don't know if I should be displaying the upload option to users
who aren't logged in (though I have a feeling I shouldnt).

Added an about page, I may want to expand that to include site operation
instructions. Kinda want some feedback as to whether we want a footer on
the site.