Modularity update
further modularized validateUser as well as offloaded all of the logic
for it into ServerLogic.Java. renamed User Dao to UserDAO and JobDao to
JobDAO, though for some reason that caused the system to stop
recognizing UserDAO, so I deleted it from the project and added it back
and then everything was fine.